Updates games and authors
