--- data/authors.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": 417,
-    "first_name": "tina",
-    "last_name": "mura"
-  },
-  {
-    "id": 857,
-    "first_name": "Moos",
-    "last_name": "Hamd"
-  },
-  {
-    "id": 94,
-    "first_name": "Tyson",
-    "last_name": "Tyt"
-  },
-  {
-    "id": 832,
-    "first_name": "tinashe",
-    "last_name": "murambinda"
-  }
-]
