+++ data/games.json
@@ -18,5 +18,10 @@
     "author": "tinashe",
     "publish_date": "2022-01-01",
     "multiplayer": "y"
+  },
+  {
+    "author": "Mike",
+    "publish_date": "1992-01-01",
+    "multiplayer": "N"
   }
 ]
